+++ lib/chewy/index.rb
@@ -10,7 +10,6 @@ require 'chewy/index/observe'
 require 'chewy/index/settings'
 require 'chewy/index/specification'
 require 'chewy/index/syncer'
-require 'chewy/index/witchcraft'
 require 'chewy/index/compiled'
 require 'chewy/index/wrapper'
 
@@ -32,7 +31,6 @@ module Chewy
     include Mapping
     include Observe
     include Crutch
-    include Witchcraft
     include Compiled
     include Wrapper